Consider max_quantity when desired exceeds OR EQUALS max_quantity
Previously, this test has a boundary condition that only triggers when
the desired quantity is some even multiple of the maximum quantity -
e.g. asking for 20, when the max is 10.

If asking for 19 (see test), we get 1x "10-pack" and thus
`sku.quantity == max_quantity`. We still want to consider the limit in
this case.